+++ packages/app/src/debug/empty/DebugNoProject.vue
@@ -4,10 +4,14 @@
     :description="t('debugPage.emptyStates.reviewRerunAndDebug')"
     :example-test-name="t('debugPage.emptyStates.noProjectTestMessage')"
     :help-link-text="t('debugPage.emptyStates.learnAboutProjectSetup')"
+    :slideshow-campaign="DEBUG_SLIDESHOW.campaigns.connectProject"
     help-link-href="https://on.cypress.io/adding-new-project"
   >
-    <template #cta>
-      <CloudConnectButton utm-medium="Debug Tab" />
+    <template #cta="slotProps">
+      <CloudConnectButton
+        utm-medium="Debug Tab"
+        :utm-content="slotProps.utmContent"
+      />
     </template>
   </DebugEmptyView>
 </template>
@@ -16,6 +20,7 @@
 import DebugEmptyView from './DebugEmptyView.vue'
 import { useI18n } from '@cy/i18n'
 import CloudConnectButton from '../../runs/CloudConnectButton.vue'
+import { DEBUG_SLIDESHOW } from '../utils/constants'
 
 const { t } = useI18n()