+++ packages/data-context/test/unit/sources/GitDataSource.spec.ts
@@ -46,12 +46,8 @@ describe('GitDataSource', () => {
     sinon.restore()
   })
 
-  // TODO: Very flaky on CI, even with a bunch of retries it still fails more often than nought
-  // see: https://github.com/cypress-io/cypress/issues/25824
+  // TODO: fix flaky test https://github.com/cypress-io/cypress/issues/23317
   it.skip(`gets correct status for files on ${os.platform()}`, async function () {
-    // TODO: fix flaky test https://github.com/cypress-io/cypress/issues/23317
-    this.retries(15)
-
     const onBranchChange = sinon.stub()
     const onGitInfoChange = sinon.stub()
     const onError = sinon.stub()