+++ packages/driver/cypress/e2e/e2e/origin/snapshots.cy.ts
@@ -74,7 +74,7 @@ describe('cy.origin - snapshots', { browser: '!webkit' }, () => {
       cy.get('a[data-cy="xhr-fetch-requests-onload"]').click()
     })
 
-    // TODO: the xhr event is showing up twice in the log, which is wrong and causing flake. skipping until: https://github.com/cypress-io/cypress/issues/23840 is addressed.
+    // TODO: fix failing test: https://github.com/cypress-io/cypress/issues/23840
     it.skip('verifies XHR requests made while a secondary origin is active eventually update with snapshots of the secondary origin', () => {
       cy.origin('http://www.foobar.com:3500', () => {
         // need to set isInteractive in the spec bridge in order to take xhr snapshots in run mode, similar to how isInteractive is set within support/defaults.js
@@ -100,7 +100,7 @@ describe('cy.origin - snapshots', { browser: '!webkit' }, () => {
       })
     })
 
-    // TODO: fix flaky test https://github.com/cypress-io/cypress/issues/23437
+    // TODO: fix failing test: https://github.com/cypress-io/cypress/issues/23840
     it.skip('verifies fetch requests made while a secondary origin is active eventually update with snapshots of the secondary origin', () => {
       cy.origin('http://www.foobar.com:3500', () => {
         // need to set isInteractive in the spec bridge in order to take xhr snapshots in run mode, similar to how isInteractive is set within support/defaults.js